ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
pillow==10.1.0
pytesseract==0.3.10
python-multipart==0.0.6
pydantic==2.5.0